Move GenderInput from name to patient
