+++ server/graphql/programs/src/queries/patient_search.rs
@@ -1,7 +1,7 @@
 use async_graphql::*;
 use chrono::NaiveDate;
 use graphql_core::{standard_graphql_error::validate_auth, ContextExt};
-use graphql_types::types::{patient::PatientNode, GenderInput};
+use graphql_types::types::patient::{GenderInput, PatientNode};
 use service::{
     auth::{Resource, ResourceAccessRequest},
     programs::patient::PatientSearch,
