+++ server/graphql/types/src/types/program/contact_trace.rs
@@ -15,10 +15,9 @@ use repository::{
     DateFilter, DatetimeFilter, EqualFilter, ProgramRow, StringFilter,
 };
 
-use crate::types::{EqualFilterGenderInput, GenderInput, GenderType};
 
 use super::{
-    document::DocumentNode, patient::PatientNode, program_enrolment::ProgramEnrolmentNode,
+    document::DocumentNode, patient::{EqualFilterGenderInput, GenderInput, PatientNode, GenderType}, program_enrolment::ProgramEnrolmentNode,
     program_node::ProgramNode,
 };
